Base modifiers
- all default modifiers except `baseUri` modifier from class `LocalModifiers` were moved to new class `BaseModifiers`
- `S3Storage` now loading `BaseModifiers`
